requirements.txt
boto3>=1.34.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
tabulate>=0.9.0
pandas>=2.0.0